# Use `map!` in `HashWithIndifferentAccess#values_at`